[RPC] Fix proposal payment timing calculation
A pre-existing block height calculation was incorrect during budget
proposal submission, which caused intermittent issues when paying out
single-payment proposals and (sometimes) the last payment of
multi-payment proposals.

This adjusts that calculation to add more consistency and reliability,
also sanitizes the input strings for the proposal name and the url to
prevent any improper JSON results.