[Security] zPIV Encryption
Currently zPIV is stored as plaintext. Created new class CZerocoinCrypter which implements the AES256CBC routine same as CCrypter, but operated over a CBigNum inplace.

This allows for CZerocoinMint to initialize and set key data for CZerocoinCrypter on construction. The cipher is derived from a pub/priv keypair that the wallet owns.