[BUG][GUI] refresh CoinControl state in CoinControlDialog
The SendWidget and ColdstakingWidget have two independent
CoinControlDialogs which share the same CoinControl object.
This creates inconsistencies, as the coin selection is carried over from
one widget to another (e.g. selecting coins in the SendWidget and then
going to the ColdstakingWidget), but some feedback is not updated in the
dialog:
- the toggle state of the "Select All" button is not changed
- the active state of btnCoinControl is not updated (this can be very
confusing for the user, as it might seem ).

We fix the first issue updating the dialog labels in the constructor of
CoinControlDialog, and the second issue checking if coinControl has any
selection (and updating btnCoinControl active state) during the show
event of each widget.