tests: make p2p_sendheaders.py a large comment
This file isn't used in our tests, so it doesn't need to pass any
future lint checks. It is also sorely out of sync with the current test
suite and needs updating before implementing it after the point when
this test would be useful, so make it's contents a large block comment
for now.