[Refactor] Return the script type from Solver
Because false is synonymous with TX_NONSTANDARD, this conveys the same
information and makes the handling explicitly based on script type,
simplifying each call site.

Prior to this change it was common for the return value to be ignored,
or for the return value and TX_NONSTANDARD to be redundantly handled.

Adaptation of btc@984d72ec659361d8c1a6f3c6864e839a807817a7 with further changes.